Conflicts:
	tajo-catalog/tajo-catalog-client/src/main/proto/CatalogProtocol.proto
	tajo-catalog/tajo-catalog-server/src/main/java/org/apache/tajo/catalog/CatalogServer.java

+++ tajo-common/src/main/java/org/apache/tajo/OverridableConf.java
@@ -178,7 +178,7 @@ public class OverridableConf extends KeyValueSet {
 
   @Override
   public float getFloat(ConfigKey key) {
-    return getLong(key, null);
+    return getFloat(key, null);
   }
 
   public void put(ConfigKey key, String val) {
